+++ src/loaders/community/standalone-loader-original.tsx
